test(dataverse): dispatch files and export private function for test
